refactor(dm): pin badge cubits' lazy:false via a shared AppShellBadgeScope (#6115)
Review follow-up on #6146: the eager-create tests pumped their own mirror
probes with a local lazy flag, so reverting `lazy: false` in main.dart
turned no test red — the flags were not actually pinned.

Extract the two badge BlocProviders + _InboxBadgeRepositorySync into
AppShellBadgeScope. main.dart and the new app_shell_badge_scope_test.dart
now pump the exact same widget, so flipping either flag back to lazy fails
the eager-create test (verified by mutating each flag independently). The
pre-fix lazy re-entrancy reproduction and the notification auth-transition
coverage are preserved; the redundant probe-based eager tests are removed.
